move proto_value tests into own file
